+++ projects/mk-kit/src/lib/components/directives/index.ts
@@ -5,3 +5,4 @@ export * from './autofocus';
 export * from './scrollspy';
 export * from './intersect';
 export * from './infinite-scroll';
+export * from './ripple';